requirements.txt
pdfplumber==0.11.9
python-bidi==0.6.7
google-genai
python-dotenv
google-cloud-aiplatform
plotly